refactor: split autounlock into new file
